# openapi.yaml
openapi: 3.1.0
info:
  title: EMR Demo API
  description: |-
    A sample API for an Electronic Medical Record (EMR) platform to demonstrate integration with mercurie.ai.
    This API exposes basic functionalities like fetching patient data, managing clinical notes, and handling encounters.
  version: 1.1.0
servers:
  - url: http://localhost:3001
    description: Local development server


# All protected endpoints will reference this security requirement.
security:
  - ApiKeyAuth: []

paths:
  /endpoints:
    get:
      summary: Get Available API Endpoints
      description: Returns a list of all available endpoints in the API.
      operationId: getEndpoints
      responses:
        '200':
          description: A JSON object mapping endpoint names to their URLs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointsResponse'

  /patients:
    get:
      summary: Get Patient List
      description: Fetches a list of all patients in the system.
      operationId: getPatients
      responses:
        '200':
          description: A list of patients.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientListResponse'

  /med-templates:
    get:
      summary: Get Drug Templates
      description: Retrieves the drug order templates
      operationId: getMedTemplates
      responses:
        '200':
          description: The encounter's clinical note.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMedicationTemplatesResponse'

  /notes:
    post:
      summary: Post a Clinical Note
      description: Creates a new clinical note or updates an existing one if an encounter_id is provided.
      operationId: postNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostNoteForm'
      responses:
        '200':
          description: Note was successfully received. The response includes the encounter ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostNoteResponse'

  /patient-summary/{patientId}:
    get:
      summary: Get Patient Summary
      description: Retrieves the clinical summary notes for a specific patient.
      operationId: getPatientSummary
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the patient.
      responses:
        '200':
          description: The patient's summary notes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSummaryResponse'
        '404':
          description: Patient not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Set Patient Summary
      description: Updates the clinical summary notes for a specific patient.
      operationId: setPatientSummary
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the patient.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSummaryRequest'
      responses:
        '200':
          description: Summary updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetSummaryResponse'
        '400':
          description: Bad request (e.g., missing summary_notes).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Patient not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /patients/{patientId}/encounters:
    get:
      summary: Get Patient Encounters
      description: Fetches the list of all encounters for a specific patient.
      operationId: getPatientEncounters
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the patient to fetch encounters for.
      responses:
        '200':
          description: A list of encounters for the patient.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterListResponse'
        '404':
          description: Patient not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /encounters/{encounterId}:
    get:
      summary: Get Encounter Note
      description: Retrieves the clinical note associated with a specific encounter.
      operationId: getEncounterNote
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the encounter to fetch the note for.
      responses:
        '200':
          description: The encounter's clinical note.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEncounterNoteResponse'
        '404':
          description: Encounter or note not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  # Describes the security scheme. We're using an API Key sent as a Bearer Token.
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API-Key
      description: "Enter your API key as a Bearer token. Example: 'Bearer your-super-secret-api-key'"

  # Reusable data structures (schemas) are defined here.
  schemas:
    # REQUEST/RESPONSE MODELS
    EndpointsResponse:
      type: object
      required: [get_patients, post_note]
      properties:
        get_patients:
          type: string
          format: url
          description: Endpoint to fetch patient list
        post_note:
          type: string
          format: url
          description: Endpoint to post notes
        patient_chart_page:
          type: string
          format: url
          description: URL of the patient chart page on the EMR, use :patientId to capture patient ID in the URL
        get_patient_summary:
          type: string
          format: url
          description: Endpoint to fetch summary notes for a patient, use :patientId to capture patient ID in the URL
        set_patient_summary:
          type: string
          format: url
          description: Endpoint to post summary notes for a patient, use :patientId to capture patient ID in the URL
        get_patient_encounters:
          type: string
          format: url
          description: Endpoint to get encounter list for a patient, use :patientId to capture patient ID in the URL 
        get_encounter_note:
          type: string
          format: url
          description: Endpoint to get notes for a particular encounter, use :encounterId to capture encounter ID in the URL
        get_medication_templates:
          type: string
          format: url
          description: Endpoint to get drug order templates

    PatientListResponse:
      type: object
      oneOf:
        - required: [patients]
          properties:
            patients:
              type: array
              items:
                $ref: '#/components/schemas/PatientDetails'
              deprecated: true
        - required: [patient_lists]
          properties:
            patient_lists:
              type: array
              items:
                type: object
                required:
                  - list_name
                  - patients
                properties:
                  list_name:
                    type: string
                    description: The name of the patient list.
                    examples: ["Today's Appointments", "My Inpatients"]
                  patients:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatientDetails'

    GetMedicationTemplatesResponse:
      type: object
      required: [med_templates]
      properties:
        med_templates:
          type: array
          items:
            $ref: '#/components/schemas/MedicationOrder'

    PostNoteForm:
      type: object
      required:
        - patient_id
        - note_title
        - medication_orders
      properties:
        patient_id:
          type: string
        encounter_id:
          type: string
          description: If present, update existing encounter. If absent, create new.
        transcript:
          type: string
          description: The trancript of the conversations from the session. Whether this field is included can be configured on Mercurie external EMR settings. 
        audio_base64:
          type: array
          items:
            type: string
            format: byte
          description: An array of base64-encoded data URLs of the audio recordings from the session. Whether this field is included can be configured on Mercurie external EMR settings. 
        note_title:
          type: string
      oneOf:
        - required: [notes_json, notes_template]
          properties:
            notes_json:
              type: object
              description: Structured notes in JSON format.
              additionalProperties: true
            notes_template:
              type: string
              description: The template used to generate the JSON notes.
        - required: [notes]
          properties:
            notes:
              type: string
              description: Unstructured notes in text/markdown format.

    PostNoteResponse:
      type: object
      required: [encounter_id]
      properties:
        encounter_id:
          type: string
          description: The encounter ID of the saved note. If your EMR doesn't support encounter listing, you may set this to "dummy".

    GetSummaryResponse:
      type: object
      required: [summary_notes]
      properties:
        summary_notes:
          type: string
          description: Summary notes in text/markdown format.

    SetSummaryRequest:
      type: object
      required: [summary_notes]
      properties:
        summary_notes:
          type: string
          description: Summary notes in text/markdown format.

    SetSummaryResponse:
      type: object
      properties:
        message:
          type: string

    EncounterListResponse:
      type: object
      required: [encounters]
      properties:
        encounters:
          type: array
          items:
            $ref: '#/components/schemas/EncounterDetails'

    GetEncounterNoteResponse:
      type: object
      required: [note]
      properties:
        note:
          type: string
          description: Previously saved notes for the encounter in text/markdown format. 

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string

    # DATA MODELS
    PatientDetails:
      type: object
      required: [id, display_name]
      properties:
        id:
          type: string
          description: Should be url-safe with at max 64 characters.
        display_name:
          type: string
        display_id:
          type: string
        display_gender:
          type: string
        display_birthdate:
          type: string
          format: date
          examples: ['1970-06-22']

    MedicationOrder:
      type: object
      required: [name, strength]
      properties:
        name:
          type: string
          description: Name of the drug.
          examples: ['Aspirin', 'Paracetamol']
        strength:
          type: string
          description: Strength of the drug.
          exmaples: ['81mg', '500mg']
      additionalProperties: true

    EncounterDetails:
      type: object
      required: [id, display_name, date]
      properties:
        id:
          type: string
          description: Should be url-safe with at max 64 characters, should be unique across patients.
        display_name:
          type: string
        date:
          type: string
          format: date
          examples: ['2025-08-10']